В product сервис были внесеные некоторые измененеие

+++ services/product/src/main/java/kz/amihady/eccomerce/product/request/CreateRequest.java
@@ -19,6 +19,6 @@ public record CreateRequest(
 
         @NotNull(message = "Количество не может быть пустым")
         @Min(value = 1, message = "Количество должно быть хотя бы 1")
-        Integer quantity
+        Long quantity
 ) {
 }
